Add GamificationToolbar UI for TrackFragment
