+++ shared/src/commonMain/kotlin/org/hyperskill/app/gamification_toolbar/injection/GamificationToolbarComponentImpl.kt
@@ -4,11 +4,10 @@ import org.hyperskill.app.core.injection.AppGraph
 import org.hyperskill.app.core.presentation.ActionDispatcherOptions
 import org.hyperskill.app.gamification_toolbar.presentation.GamificationToolbarActionDispatcher
 import org.hyperskill.app.gamification_toolbar.presentation.reducer.GamificationToolbarReducer
-import org.hyperskill.app.gamification_toolbar.presentation.reducer.GamificationToolbarReducerImpl
 
 class GamificationToolbarComponentImpl(private val appGraph: AppGraph) : GamificationToolbarComponent {
     override val gamificationToolbarReducer: GamificationToolbarReducer
-        get() = GamificationToolbarReducerImpl()
+        get() = GamificationToolbarReducer()
 
     override val gamificationToolbarActionDispatcher: GamificationToolbarActionDispatcher
         get() = GamificationToolbarActionDispatcher(
